requirements.txt
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
coverage==7.0.5
exceptiongroup==1.1.0
Flask==2.2.2
greenlet==2.0.1
idna==3.4
importlib-metadata==6.0.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.0
pluggy==1.0.0
PyMySQL==1.0.2
pyodbc==4.0.35
pytest==7.2.1
python-decouple==3.6
requests==2.28.2
SQLAlchemy==1.4.46
tomli==2.0.1
urllib3==1.26.14
Werkzeug==2.2.2
zipp==3.11.0
